refactor: extract source filtering functions
